feat: implementa PedidoCompraItemService

+++ src/main/java/com/gestaooserp/dev/repository/PedidoCompraItemRepository.java
@@ -3,5 +3,5 @@ package com.gestaooserp.dev.repository;
 import com.gestaooserp.dev.entity.PedidoCompraItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PedidoCompraItemRepository extends JpaRepository<PedidoCompraItem,Integer> {
+public interface PedidoCompraItemRepository extends JpaRepository<PedidoCompraItem,Long> {
 }
